Updated openstack/openstack Project: openstack/neutron-vpnaas  bc2a1d775408a630e483794437acb90bd4a3c3c9
VPNaaS: Reorganize test tree

Similar to what is happening in Neutron, this commit renames and moves unit
test files so that they match the directory structure of the code tree.
Since the VPN repo's test files have imports from Neutron, there are import
changes as well, to the imports to match the Neutron test file organization.

A check script is added to monitor the test file naming and locations (to
ensure consistency).

Change-Id: I875285230127ed660137d271f7e98d82eb57ffab
Partial-Bug: #1440834