Updated openstack/openstack Project: openstack/sahara-image-elements  abcc1c23e7ed9b898e3e90c2d6480747e5a019b7
mysql: use install-packages and pkg-map elements

Make use of the install-packages element, so it is easier to specify the
packages using a declarative yaml file.
To keep the different list of packages depending on the distro, make use
of the pkg-map element so the right packages per-family and per-distro
(Fedora, in this case) can be chosen.

Change-Id: I64797fd075204609d54dbd01c896acb7ab638a29