Updated openstack/openstack Project: openstack/cinder  ea109b5f24dca93fd6f660bc436a685d6101bcea
Include boot properties from glance v2 images

In order for users to take advantage of COW volumes created from
a glance image, Cinder must be configured to use Glance API version
2 (default is 1). In version 2, the required boot metadata (kernel_id
and ramdisk_id) are no long stored in the 'properties' dict, but as
standalone fields in the GET response from glance. The existing cinder
parser for the glance request is not aware of this and the volume
created form a v2 image will lack this required metadata.

This was causing the recent Ceph CI gate failures for
test_volume_boot_pattern.

Change-Id: I688898b3841691369d73887f7eabdceb05155db1
Closes-Bug: #1439371