Updated openstack/openstack Project: openstack-infra/puppet-zanata  aa25be34df9efdf862038657ff8c6edbaf51f928
Add OpenID login provider support to Zanata config

We want users to sign in using openstackid.org. This requires
changes to the Zanata config to specify an OpenID provider and
to disable the existing internal login so we don't wind up with
multiple login buttons. Make OpenID configuration optional so
others can still use the module without it. Also allow the admin
user list to be configured.

Change-Id: I3688688ef37fea602b107fba17dcc8d6f40f9275