Updated openstack/openstack Project: openstack/heat-translator  d4fc768ada9981f4b915bc167929ee1c6a29d850
Follow the hosting chain to find the host server

Since a hosting chain may exist among several node templates (e.g.
a web application hosted on an application server, the application
server hosted on a web server, the web server hosted on a compute
node) update the code to follow the chain and return the actual
compute node host. The unit test for this patch will be covered by
bug #1435545.

Change-Id: I01bab61bc45b4031cd106295d216982332361b6b
Closes-Bug: #1439444
Partial-Bug: #1435545