Updated openstack/openstack Project: openstack/governance  d6a3dd7b99356a5350e30d7c44f932f583f01d1c
Add a two-organization rule to the diverse-affiliation tag

Thierry suggested the addition of a two-organization rule during the
review of the team:diverse-affiliation tag.  This patch adds that rule
and notes that it does not trigger any changes to the application of
the tag at this point.  However, it comes close for Trove.

This patch also includes a wording change from "company" to
"organization" in the tag definition, which was suggested by Anne
during the original review.

Change-Id: I132aa31a2c33580685619c41ec2d7d7a84d858c5