Updated openstack/openstack Project: openstack/governance  8d6fc8011313cdf27c9c272c1b8d8ad3b830e270
Add "release:has-stable-branches" tag

This tag lets us differentiate between projects that happen
to have stable branch cut out of their last release in the
development cycle (like Oslo libraries or Python client libs)
and projects that consciously make a release at the very end
of the 6-month development cycle to match with the "managed"
coordinated release (like Swift or Manila).

Those are different bits of information to convey to our users
(will they have a release to match with the coordinated 6-month
release, vs. will they have a stable branch to draw
backward-compatible safe updates from).

To make it easier to judge what that means, this change directly
applies the new tag to the existing set of projects.

I also clarified the meaning of the "release:at-6mo-cycle-end" tag
and removed it from projects that don't make a coordinated release.

Change-Id: Icb3b5a348e0a45437d6ac6a044ed098c6380ee97