Updated openstack/openstack Project: openstack/nova  0f70e1035ff2d5ca9ba4df88b8c14fb54e9db6a3
remove unneeded teardown related code

The DatabaseFixture ensures we have a blank database at the beginning
of a test, previously there was a teardown method here, which was
deleted. However there remains a bunch of state tracking which was
related to that teardown. Cleanup up all the extra variable references
to avoid future confusion with this test.

Change-Id: I5527413dfcc6b9e0691bfc229247e7707b75cb45