Updated openstack/openstack Project: openstack/tempest  040fc6c6e7ac77b050d640a37589384cc220a0d6
Use network from get_tenant_network in scenario

In the scenario tests create_server helper method the returned network
from get_tenant_network() was not being used it was returned and the
method to add it to the kwargs dict was called, but the resulting
updated dict was never used. This patch corrects the oversight by
updating the kwargs dict to ensure we actually pass the right network
into the create server call.

Change-Id: I1976ed6eb2edfa77971bd908fe47514c1a801b98