Updated openstack/openstack Project: openstack/nova  c0c7d2bd6cc726278d9a6da6deb0ce203407646e
Fix Enum field, which allows unrestricted values

For some reason, the Enum field allows valid_values to be None, which disables
all the is-this-a-valid-value checking, thus defeating the entire purpose of
the type. This patch removes that possibility so that we don't let anything
creep into the RPC-affecting API that would be unrestricted.

The change that introduced this was I7848f8b934c169a76b1cfd875f7905273f9f914b

Closes-Bug: #1441243
Change-Id: Ia33994402cca5b4569d72860ab5da675134f5a95