Updated openstack/openstack Project: openstack/taskflow  5ef3e6d2e9d82f8e569154c4d1b79e4ec7f0eb59
Allow providing a node stringify function to tree pformat

Allow providing a node to string function that will be activated
when the root node pformat routine needs to format a node into
a string (for viewing purposes); this allows for subclasses or
other users of nodes to provide there own function that they can
specialize as they see fit (without duplicating the pformat
routine itself).

To start use it to print out the node link target in the fake
in-memory filesystem when a link is encountered.

Change-Id: I33b8f1ec8d30d72ae35971e35f5edd84a9145db7