Updated openstack/openstack Project: openstack/ironic  bb75894ba2de508a9eab9ac766f16d8f23bf3a50
Fix VM stuck when deploying with pxe_ssh + local boot

When deploying a machine with the DIB ramdisk, pxe_ssh driver and
local boot the VM will keep rebooting forever after the deployment is
completed. This happens because Ironic will change the boot device from
PXE to DISK as part of the deployment but the change will only take effect
after the VM is powered off and on again (that's a virsh problem). This
patch is fixing this problem by making Ironic reboot the machine after
the deployment is completed (reboot in Ironic means power it off and
on again).

Closes-Bug: #1439664
Change-Id: I4aa859b00a8bba186e920e09734ec2e4c45dec3e