Updated openstack/openstack Project: openstack/python-keystoneclient  53dc900025c34378854425f40e3dea82d0416254
Return None for missing trust_id in fixture

If the trust_id is unset it raises a KeyError. This is unusual from a
python perspective (if nothing else it should be AttributeError) and
different to all the other attributes of the fixture.

Return None if no trust_id is set on the fixture.

Change-Id: I15d33d77027a188fa47df18387c4610908f8e2d2