Updated openstack/openstack Project: openstack/openstack-manuals  66c6c2a5907a42396a743e46bed49c606a585b2a
Update links in Community Support Appendix

Use link without content for User Guides to work with both old and new
guides.

Remove API links, they are all now on developer.openstack.org - and the
first link in the section contains all documents, so we do not need to
duplicate the information here.

Change-Id: I1e6497c3f8a4dd5b647a306964f7def779cd85bf