Updated openstack/openstack Project: openstack/python-keystoneclient  045c9dab6dfd599de42a7a7ac4a677e1a793b67f
Provide a generic auth plugin loader

For keystonemiddleware, shade and other projects that do more
complicated option loading than simply CLI or CONF file provide a means
to load an auth plugin where options are discovered by a provided
function.

This plugin is designed to work with the options as provided by
get_options rather than either the argparse or CONF registration
functions.

Use these as the default loading mechanism for the existing argparse and
CONF functions as it standardizes the mechanism between the two sources.

Change-Id: I15634ac30581c7aea14e709f12fb202570190f46
Closes-Bug: #1428900