Updated openstack/openstack Project: openstack/cinder  a70b3c53484b6555fffb75cf55af931ed3b8449d
Correct cinder hacking check numbering

We have a couple of hacking checks that are specific to
Cinder that were written a while back.  Unfortunately, when
they were written the numbering scheme for hacking checks was
not understood.  We used N3xx when we should have used C3xx.

This patch corrects that mistake.

Change-Id: Ia17797005d444ab53a45b47b80b97799114001ee
Closes-bug: 1440833