Updated openstack/openstack Project: openstack/python-keystoneclient  655719d6c96bc473c03a4f51fbb12160f80bfcb4
Removes discover from test-reqs

The discover package is part of Python 2.7, so we don't need
to explicitely require it. It's by the way annoying for
distribution package maintainers, as the file has to be
patched.

Keystone team has decided to drop python 2.6 support[1],
so, it's safe to remove it from test-requirements.txt.

[1] http://eavesdrop.openstack.org/meetings/keystone/2015/keystone.2015-11-24-17.59.html

Related-Bug: #1519449
Co-Authored-By: Thomas Goirand <thomas@goirand.fr>
Change-Id: I5bb258a4b305c0084be50bc22fe7a0e6a4f65aad