Updated openstack/openstack Project: openstack/cinder  580997f6ad242caca79c8760eecb1904ca2185cd
Rbd update volume stats in wrong way

Cinder volume uses a RADOS pool to store volumes and
update storage stats periodically. However, rbd driver
reports the whole cluster stats but not the pool's. This
is wrong. The right way is to report the pool stats but
not the whole cluster.

Change-Id: Ie18324ebd8b85f7b63dc034ba20857ae9e470804
Closes-Bug: #1434441