Updated openstack/openstack Project: openstack/cinder  fffd9a057ba107108513916b26205eab66b7dc26
Delete the temporary volume if migration fails

Issues resolved in this patch include the following changes:
* A temporary volume is created on the destination host before migrating
the data from the source. However, if the creation of this volume fails,
its record will be left in the database as redundant information. This
patch will remove the database record if the creation fails.

* If attaching the remote dest volme fails at initialize_connection
due to timeout, we need to terminate the connection. Otherwise, the dest
volume will not be released and successfully deleted.

Closes-bug: #1398177
Change-Id: I5bc75bd7652841fddd66481ee001fb682168215c