Updated openstack/openstack Project: openstack/nova  b3c2b0cc6d4748f08ae2814f634411f9dde1e06d
Fix multipath device discovery when UFN is enabled.

This currently returns an invalid path of `/dev/mapper/${WWID}`
when UFN is enabled leading to failures later on when we attempt to
use the volume.

The output of `multipath -l /dev/${path_device}` should always list
the correct device identifier to use with this path as the first
word on the first line.

Closes-Bug: 1401799
Change-Id: I07957fe43e68a55ece10343a6cee83a9ab7148a8