Updated openstack/openstack Project: openstack/heat-translator  17afcb04d18a982b512b0cd428e04dec35fca99e
Fix coverage so that it works for this repo

PBR provides test and testr entrypoints, the latter of which is not
recommended as its serviced directly by testr. Switch to test in
tox.ini and pass an option to PBR indicating the correct package
name to use when calling coverage. Also update .coveragerc with the
corrected source path, and stop omitting the tests themselves (after
all, tests should cover themselves!).

Change-Id: I04b1a5393979c5d82a794d35189743f688bdfdd8