Updated openstack/openstack Project: openstack/keystone-specs  46676971fb3e9b3bca9bf3e524c07fdb70c524c7
Update path for listing a project's endpoint groups

The original route didn't exist in any version of keystone yet,
so the API was broken in that regard. Also, it should be changed
to be more RESTful; the list of resources that is returned should
be the last argument of the path.

Change-Id: I8b6a0713cf01a6c0107ca82fb35e99c42f0346a5
Related-Bug: 1421968