Updated openstack/openstack Project: openstack/ironic  f7786d92b77daa8acb5b7047a3784cbb3f57f59a
Fix mocks not being stopped as intended

Mock objects were being created that were intended to be started and
stopped.  But they were not actually being stopped.  Fix it so they are
being stopped correctly, add autospec=True to the mock objects, and
update some side_effects to work with autospec=True.

Change-Id: I4ced8eb8a4fe874405f0b53132a38e276eef01b5