Updated openstack/openstack Project: openstack/ironic  55a775182f1e40032901a3ed4c78abc750b81c41
Add retry logic to _exec_ipmitool

When using bridged IPMI buses, the possibility exists for the BMC
to return messages indicating that it is too busy to service the
request at that moment.  This results in ipmitool returning an
error to the user instead of retrying the operation.

As a result of the error being returned, the current operation
being attempted by ironic is failed and is never retried as the
present retry configuration is purely a pass-through to ipmitool.

Added logic to allow for the identification of errors returned by
ipmitool that can be retried coupled with logic to retry the failed
operation until the configured retry window has passed.

Included three tests which are geared to exercise the retry logic
and ensure that the retry logic works as expected, including that
the retry operations are only performed a specific number of times
based on the conditions.

Updated one of the pre-existing tests that tested _exec_ipmitool
act of returning an exception to ensure that the retry logic is
executed only once.

Change-Id: I698a8ff1cb70d7758ed05e59b858d9e40708dd05
Closes-Bug: 1431929