Updated openstack/openstack Project: openstack/cinder  f07713139394758f0734055b5a3389a8a13538cf
Fix multipath device discovery when UFN is enabled.

This currently returns an invalid path of `/dev/mapper/${WWID}`
when UFN is enabled leading to failures later on when we attempt to
use the device.

The output of `multipath -l ${path}` or `multipath -l ${wwid}`
should always list the correct device identifier to use with this
path as the first word on the first line.

Closes-Bug: 1401799
Change-Id: Ib371b699fadcbbbb666e08eb0124c442e94a55e8