Updated openstack/openstack Project: openstack/tempest  839567cc0bea198d9cb1412bb39f3c617199dcf9
Skip test_volume_boot_pattern until bug 1439371 is fixed

The check-tempest-dsvm-full-ceph job sets volume.storage_protocol to
'ceph' in devstack-gate and the test is failing with a ceph backend, so
skip it while that's being debugged.

The plan is to get the ceph job passing consistently so we can make it
voting on nova/cinder/glance changes so we can tell when a change
regresses it.

Related-Bug: #1439371

Change-Id: Ia1cba598a8243dd4a48613c870a25de578b79fe4