Updated openstack/openstack Project: openstack/nova  3d31f6684a78539f5545fcabed6c2b96cb639459
Merge evacuate functional tests between v2 and v2.1

Currently v2 and v2.1 have separate functional tests and their
corresponding sample files. As v2 and v2.1 are supposed to be identical,
there is overhead to maintain two set of functional tests and sample files.
We can have one set of tests which can run for both v2 and v2.1.

This commit merges evacuate functional tests.

V2 has below 2 extensions for evacuate
- os-evacuate
- os-extended-evacuate-find-host
which are merged in single plugin 'evacuate' in v2.1

In V2.1 os-access-ip is separate plugin where in v2 it was mixed in server
plugin itself. So os-access-ip extension is enabled for server tests to run
merged functional tests for both v2 and v2.1.

Change-Id: Ib3c7fd1ecc9bc9a188350a0ec42299fc01919445