Updated openstack/openstack Project: openstack/manila  9a2f122d41f41c1e82491ff328ec6fd1f24b8426
NetApp cDOT driver uses deprecated APIs for NFS exports

The nfs-exportfs-* APIs that the NetApp cDOT driver currently uses
for NFS export management are 7-mode APIs which were left in for
backwards compatibility. It's not the recommended/supported way
to do NFS export management. We need to modify the driver to use
the supported set of export-policy-* and export-rule-* APIs.

Closes-Bug: #1370761
Closes-Bug: #1437509
Change-Id: I91347d27ba69d4a20fe73ce7e75f58c9818ea6d8