Updated openstack/openstack Project: openstack/nova  69ea06a613a5b44bab747379ac757681a1579bff
Manage compute node that exposes no pci devices

Booting a VM with a pci device causes an exception when the host doesnt
have any assignable pci devices.
If a host doesnt expose any pci devices to nova, its pci_stats are set
to None. This causes an exception when the PciPassthroughFilter checks
if the host can support the pci requests.

Closes-Bug: #1435981
Change-Id: Ia9fbec9c7a3afa9b7ea73c43b21db0909a195a36