Updated openstack/openstack Project: openstack/ceilometer  21216a297da754df2632dbaa2af264a01e3af3ce
Avoid conflict with existing gnocchi_url conf value

If for some reason there is an existing ceilometer.conf lying around
that has a gnocchi_url setting in the [alarms] section then that
value will be used when running the test. We need to use that value
when validating the match, not a hardcode host string.

Change-Id: I0e23e3c380543fb491468977294cdbe11f58fa07