Updated openstack/openstack Project: openstack/cinder  672a1d4362268cd614adb2b424c982d6555d1629
Create initiator id if not exist in VMAX driver

This problem was discovered in a customer's environment. If the
initiator id (hardware id SMI-S instance) does not exist on the
array, the attach volume operation will fail. In this patch,
the initiator id will be created if it doesn't exist.

Change-Id: I868118528d831a9f49081a6393e64eb7a414b36f
Closes-Bug: #1440427