Updated openstack/openstack Project: openstack/nova  5240068c90e93ab1d57aea62680fa888be43b8d5
websocketproxy: Make protocol validation use connection_info

With this change, we avoid the need to have config '*_baseurl' config
options correctly set up on nodes running console-proxy services, as the
access_url is now available after token validation.

UpgradeImpact: Websocket proxies need to be upgraded in a lockstep
with the API nodes up to this commit (or when upgrading to Kilo),
as older API nodes will not be sending the access_url when authorizing
console access, and newer proxy services (this commit and onward) would
fail to authorize such requests.

Change-Id: I721bca407bf9d3fb33a3461c04d392284448d4a6
Closes-bug: #1442048