Updated openstack/openstack Project: openstack/nova  74ca660ab688e15ccd59ddfbfcdc9e1cecdc553d
Fix how service objects are looked up for Cells

When querying for services in the hypervisor API extension the queries
need to be proxied through to the child cells.

Note: We need to proxy the calls thru the HostAPI in order to have the
same method for both non-cells and cells.

Co-Authored-By: Andrew Laski <andrew.laski@rackspace.com>

Change-Id: I3ff3976d24aef108a990fb8ff64afd29c96a1e2f
Closes-Bug: 1427351