Updated openstack/openstack Project: openstack/tempest  bf804da849a5466de8b9a5f7f9b90a1df07012e4
Remove EC2 test case that incompatible with Amazon

This test tries to allocate address and disassociate it without association.
Amazon allows to do it for EC2-classic and does not throw error.
But EC2 implementation in Nova throws error.
We have the same test in our own test suite against stackforge/ec2-api
(but it's not merged yet) and I checked it against Amazon -
https://review.openstack.org/#/c/171050/

I suggest to remove this test from tempest as incompatible with Amazon.
Also because of EC2 is deprecated in nova and can't be fixed.

Change-Id: I711314408ec7c4ea75f93348a7335843713e9a2b