Updated openstack/openstack Project: openstack/nova  77d19b97c69d48ef2512fb3d14b0486ed2137b39
Honor uuid parameter passed to nova-network create

The nova api for creating nova-network networks has an optional
request parameter "id" which maps to the string uuid for the
network to create. The nova-manage network create command represents
it as the option --uuid. The parameter is currently being ignored
by the nova-network manager. This change sets the uuid when creating
the network if it has been specified.

Closes-Bug: #1441931

Change-Id: Ib29e632b09905f557a7a6910d58207ed91cdc047