Updated openstack/openstack Project: openstack/requirements  05d010e5e2f90443d9b09feb5c9e7832cc9160a5
Add pyngus to requirements

oslo.messaging use pyngus for the one of its functionnal job (amqp1).

This python module is not in the requirement of oslo.messaging
because it can't be build on ubuntu trusty due to a missing c library
qpid-proton, and because the driver is experimental and can be run only
on fedora >= 21 for now, so the installation of it have been hardcoded
in devstack for fedora to allow gating on this driver.

Since all python modules installed via pip are now checked against
global-requirements we must have this deps here.

This change does that.

Closes-bug: #1442488

Change-Id: I91a7655becc4d559e3928b67cf70586cdaf42a29