Updated openstack/openstack Project: openstack/nova  f9347ebd3ad8d554af12ec256b5603d6ab001fc7
Scheduler: deprecate mute_weight_value option on weigher

The nova.cells.weights.mute_child.MuteChildWeigher used a large
mute_weight_value flag to inflate the weigher's value. This was needed
before weight normalization was in place, so that the weigher could win
against other weighers. Now, with weight normalization in place this is
not needed any more, since the maximum value that a weigher would take will
be 1.0 (weights are normalized between 0.0 and 1.0). We should use the
multiplier to get the same behavior (i.e. make the weigher prevail
against other weighers), the multiplier has been also increased so
that it actually mutes the other weighers an does not break backwards
compatibility.

DocImpact: The mute_weight_value has been marked as deprecated so the
operators need to adjust the mute_weight_multiplier to a proper value if
they weren't using the defaults.

Closes-Bug: #1340068
Change-Id: Idebd7649909567034b01ce04416ce315fb936b17