Updated openstack/openstack Project: openstack-infra/project-config  439a7c034a0aecc53e203d8def67bee8c6f7829b
Change proposals fall back on proposed branches

During the RC period for a project, there may already be a stable
branch for the release on the triggering project and a proposed
branch on the target project. If so, we want stable changes proposed
to the proposed branch. That's not confusing!

Change-Id: I488bb9ff63daf9d012e41dbaa178b1656eaeffad