Updated openstack/openstack Project: openstack/manila  ac5f06e01148ded83935c583ff66c2f5f3265ca5
Mock out base share driver __init__ in EMC driver

Two EMC driver unit tests failed because it failed to load Neutron
network plugins. This is because stevedore.extension.ExtensionManager
is mocked to return [] for the EMC parent driver (EMC parent driver
uses stevedore framework to load backend plugins such as VNX and
Isilon), however, the base share driver tries to load Neutron network
plugins and expects the list to be non-empty.

This patch mocks out the base share driver __init__ method so that
it will not try to initialize the base class and will not try to
load the Neutron network plugin.

Change-Id: I57ff71f638b48c1485b32401addc909002fdc768
Closes-Bug: #1442788