Updated openstack/openstack Project: openstack/keystone  0325fdf92b98764900cbd926c24747b18e99cc96
Refactor MemcachedBackend to not be a Manager

The MemcachedBackend class doesn't act like a normal Manager since
it doesn't use the superclass to import the driver. All it did is
forward item references to the driver, but this is obvious and easy
enough to implement locally. Making MemcachedBackend not subclass
from Manager makes subclasses of Manager consistent.

bp stevedore

Change-Id: I1d865220e8225bf7a9a1b016eea1d364c70ed832