Updated openstack/openstack Project: openstack/manila  7df3c008a8cdf5a834ae8293a33f85f9459d4e98
glusterfs_native: Fix Gluster command call

In the delete_snapshot API, the Gluster command issued to delete the
snapshot ends up being run in the interactive mode of the
Gluster-CLI. So instead of an XML output, the command results in a
Gluster-CLI read error. Fix this by forcing the Gluster command to be
run in the script mode.

Change-Id: Ic09f59732bf08942a9d216f70f1fc969d9ae0a2d
Closes-Bug: #1442339