Updated openstack/openstack Project: openstack/barbican  4f7c5f6b465b18c5eabf42148a2c93512450d284
Add Bandit security static analysis checking via tox

The plan is to add Bandit to the commit gate.  Bandit is
produced by the OpenStack security team to do static code
analysis for security implementation failures.  It tests
the code base for ~30 common security problems in Python
with more to come.  The bandit.yaml configuration file
is set to the value recommended by the Bandit team.

run with:
    tox -e bandit

Change-Id: I46cde803119d0162b6256346e423079cadac17cb
Partially-Implements: blueprint gate-bandit