Updated openstack/openstack Project: openstack/cinder  cade28c2c8fbd1b241569874abe6a50012ef0039
VNX Cinder Driver should report 0 free_capacity_gb in some scenarios

When the storage pool is Initializing, Offline or Deleting, no more LUNs
can be created before the pool gets out of the state.
So when a pool is in the 3 states, its free capacity is de facto 0.

This patch is to add logic to report 0 free capacity accordingly.

Change-Id: I4ef56d48e987795a48e6dcb23615e4b2a7580452
Closes-Bug: #1431233