Updated openstack/openstack Project: openstack/neutron  7c46e082cdb6055660a004defcfe7abe4a343fb6
Handle non-index lookups in native OVSDB backend

ovs-vsctl get/set/clear/list can use a record_id that is not an
index on the table being queried. For example, the Controller table
can be queried by a bridge name.

This patch implements the lookup table that ovs-vsctl uses to do
these lookups.

Change-Id: I1983c48c5839df016046ba2596c7c4affa1ebe00
Closes-Bug: 1435567