Updated openstack/openstack Project: openstack-infra/project-config  f246cdd164df272689e22b94755da7f03e172535
Add check-tempest-dsvm-f21 to experimental queue for openstack/nova

Live snapshot with the nova libvirt driver is disabled by default
because we know that it race fails in the gate with libvirt 1.2.2 on
Ubuntu Trusty 14.04.

The Fedora 21 job runs with libvirt-1.2.9.2 and qemu-kvm-2.1.3 so if we
need to test experimental features in the libvirt driver that are
enabled conditionally based on versions of libvirt/qemu we're testing
against, we should be able to use the Fedora 21 job in the experimental
queue for that now.

Related-Bug: #1334398

Change-Id: Idd7cca3b77147be6583c78c6a811a5f411ad5bf4