Updated openstack/openstack Project: openstack/training-guides  37235df89e8e3c6cb40db98c44722b8312cbcdb7
Adds upstream training to OpenStack gate

Update the current gate 'slides' to build and publish upstream
training content. This patch does the following changes.

* Adds generateslides script to use sphinx/hieroglyph to build slides
* Updates tox.ini to use generateslides script
* Update .gitignore to ignore _build/ directory which is the default
  folder for generated output generated by sphix.

Hence forth, any changes to the upstream training rst slides will be
gated and published, also it is essential to note that these jobs are
new and may not be accurate or reliable to address all the errors in
the rst slides.

Change-Id: I8a220dd31501f9b3733a6282d7048fe283d2a0d1