Updated openstack/openstack Project: openstack-infra/project-config  b299b44595d29ae75bedbf2bc9783a7b8398d193
Enable a check job for networking-ovn

This enables a simple check job for the stackforge/networking-ovn
project. Initially, this job simply verifies we can bring devstack
up with OVN configured as the Neutron ML2 MechanismDriver. As OVN
functionality grows, we'll slowly enable running tempest tests.

Note, the job is started as non-voting initially until we can
ensure it's stable. We'll move it to voting once we can ensure it
won't block current development and it's stable.

Change-Id: I14866fb86f6873e59ee3f9f0b6e280a235248fc9