Updated openstack/openstack Project: openstack-infra/tripleo-ci  7a2c3d913d2f41034b2ae53563020faa349cc817
Pin qemu-img on Fedora 21 jobs.

There was a recent regression due to a security fix
for the qemu-img package on Fedora 21. Similar
to this issue we had previously:
https://bugzilla.redhat.com/show_bug.cgi?id=1200043

Until the issue is resolved we want to pin to the earlier
version for CI.

This also pulls in a required update to DIB to
make the package updating happen at a more
intuitive time.

Change-Id: I7567dc4f39a29423435edf441264cf72262593e1